Add a task addition function to the project
Two functions were created, one of which changes the active project, and the other adds a new to-do object to it. The Todo class is now imported into the projects module instead of the function. The task add function now creates a "to do" object with this class, which is added to the active project.