Change project h1 elements to buttons
The projects to be listed were changed to buttons, as events will be added to them, and their style was slightly modified.
asideUpdateDOM was changed to update myProjectsDOM, as it contains the list of projects that we want to update.
In addition, a small change was made to the active project switch function, where it no longer takes a projects array as a parameter.