Fix an issue with an active project
When the user pressed the project button, the active project did not change because it called a variable instead of a function. The variable only held a copy of the currently active project, and did not update the value when the button was pressed. Now the function always returns the updated value.