Expand task card view
An event was created to toggle the view of the task card. The DOM content was created in the card, which was given the class name "hide", and the CSS file was defined with this class name display: "none".

A separate function was created to change the state of this class selector with the toggle() function.

The "AddGlobalEventListener()" function was also created in the Utils module, which can be used to conveniently create new events. This function was also used to create the task card extension toggle event.