Fix project deletion issue
The problem was that when all projects were deleted, the name of the last project remained visible, along with the delete and add task buttons. Now, if there are no projects left, the buttons are removed and the project title shows the text "No Projects".

When there are no projects left, and the user creates a new one, it is automatically updated to the active project.