Add a delete task button
The delete task button is in a hidden section of the task card. If the user opens additional information from the details button, a delete button will appear, which he can use to delete the task from the project.

A couple of helper functions had to be created for deletion. First, a function that removes the task from the activeProject's tasks array. The second function was already created, which can hide the DOM element, and had to be given the created wrapper div as an argument.