Convert task modal from DOM to HTML
The task modal was moved to the HTML file because it makes adding and editing tasks much simpler. For this reason, the handling of the modal events was moved to the modals.js module and the task modal creation module itself was removed.

The task add button itself is now added to the main-project-DOM.js module. In addition, the button is now rendered in the same function as the project delete button.