Add task execution logic
A checkbox marks a task as completed. Tasks were given a completed status, the date the task was completed, and a reminder. In addition, methods were added to mutate state of properties.

A reminder had to be added to the form, and all functions that required object properties had to be updated.

Completed tasks and uncompleted tasks are placed in their own DOM containers.