Fix displaying containers
Tasks were divided into two containers: today's tasks and upcoming ones. The problem was that when the user deleted all tasks, the container would still be visible or the other one would not be visible at all. Now the function first checks if there are any tasks left, after which it optionally adds the containers to the page.