Move DOM content to the same module
The content of the Aside.js module was moved to the same module as the main DOM. Due to the size of the project, keeping the DOM content in the same module makes maintenance and DOM manipulation simpler.
The card creation function was split into two halves, one that creates the card and the other that updates the cards.