Add local storage
The project now saves changes and loads them when the page is loaded. A separate module was created for local storage, which contains functions for saving and loading.

For this, some changes had to be made to the project and todo classes. The constructor now takes all values, and some default values. The created instances were updated in the modal and the default project.

In addition, a helper function had to be created for updating projects after projects have been loaded from local storage.