Change the layout of the task card extension
The layout of the card extension was changed from grid to flex, because the extension, which contained a lot of text with grid, contained a lot of unnecessary white space.

An animation was also added to the card opening, and instead of controlling the card extension with display, opacity and max height were a better option.