Style aside
The My projects title was removed and replaced with the Add Project button from the header. SVG icons were added to the Add Project and Project buttons. The buttons also had a zoom animation on hover.

The Aside element was moved to the HTML file, instead of being created using the DOM.