Change the extension button to an SVG icon
The expansion button is now a typical expansion arrow/triangle SVG icon.

To do this, had to change the function that defined the content of the button. Instead of changing the text of the button, it now changes the source of the SVG icon. The SVG icon is either an up or down arrow.