Fix default project loading in local storage
If local storage did not contain data, it did not load the default project correctly. To fix this, the loading process had to check if there was data in local storage, if not, then the defaultProject() function was loaded as data.
In addition to these, the modal functions also had to be fixed, because if there were no projects on the page, the events for the buttons caused an error because they had not been created yet. Now the functions are only run after the buttons are created.