Fix task appearing on page without projects
When there was one project left on the page and a task was added to it before the project was deleted, this task remained visible on the page. The problem was the updateTaskLists() function that updated the DOM containers only when an active project was found, so it did not run because no active project was found. Now if the project is not found, it automatically clears the DOM containers.