Add a task due date counter
The number of tasks for the current day is displayed on the edge of each project. For this, a new method was added to the project class that counts the tasks based on the current day.

A function was created for the DOM element that displays either a checkmark (tasks completed) or a count of uncompleted tasks.

The updating of various functions had to be changed so that they also update the task counter.