Confirm task deletion with modal
The user is asked if they really want to delete the task. For this purpose, a modal was created and its events were moved from the event.js module to the modals.js module.
The style and animations of the delete button were modified to be similar to the other buttons. The icon for the delete task button is the same trash can as the delete project button.