Remove unnecessary functions
There were far too many functions for updating the DOM, so these were reduced to a few. First, the updateMain() function was renamed to updateTaskCards() and its purpose is only to call functions that update the content of task cards and their containers. The DOM update of the active project was moved to the updateDOM() function.